// _examples/node-ts-header/example.gen.ts
export interface ExampleServiceClient {
  /**
   * Simple method, no headers
   */
  ping(headers?: object, signal?: AbortSignal): Promise<PingResponse>

  /**
   * Single header field (authToken sent as HTTP header, userID in body)
   */
  getUser(req: GetUserRequest, headers?: object, signal?: AbortSignal): Promise<GetUserResponse>

  /**
   * Multiple header fields (authToken and role as headers, username in body)
   */
  createUser(req: CreateUserRequest, headers?: object, signal?: AbortSignal): Promise<CreateUserResponse>

  /**
   * All fields as headers (no body)
   */
  deleteUser(req: DeleteUserRequest, headers?: object, signal?: AbortSignal): Promise<void>
}

//
// Server interface
//

export interface ExampleServiceServer<Context = unknown> {
  ping(ctx: Context, req: PingRequest): Promise<PingResponse>
  getUser(ctx: Context, req: GetUserRequest): Promise<GetUserResponse>
  createUser(ctx: Context, req: CreateUserRequest): Promise<CreateUserResponse>
  deleteUser(ctx: Context, req: DeleteUserRequest): Promise<void>
}

//
// Schema types
//

export interface User {
  id: number
  username: string
}

export interface GetUserRequest {
  authToken: string
  userID: number
}

export interface GetUserResponse {
  user: User
}

export interface CreateUserRequest {
  authToken: string
  role: string
  username: string
}

export interface CreateUserResponse {
  user: User
}

export interface DeleteUserRequest {
  authToken: string
  userID: number
}

export interface PingRequest {
}

export interface PingResponse {
}






//
// Client
//

export class ExampleService implements ExampleServiceClient {
  protected hostname: string
  protected fetch: Fetch
  protected path = '/rpc/ExampleService/'

  constructor(hostname: string, fetch: Fetch) {
    this.hostname = hostname.replace(/\/*$/, '')
    this.fetch = (input: RequestInfo, init?: RequestInit) => fetch(input, init)
  }

  private url(name: string): string {
    return this.hostname + this.path + name
  }

  queryKey = {
    ping: () => ['ExampleService', 'ping'] as const,
    getUser: (req: GetUserRequest) => ['ExampleService', 'getUser', req] as const,
    createUser: (req: CreateUserRequest) => ['ExampleService', 'createUser', req] as const,
    deleteUser: (req: DeleteUserRequest) => ['ExampleService', 'deleteUser', req] as const,
  }

  ping = (headers?: object, signal?: AbortSignal): Promise<PingResponse> => {
    return this.fetch(
      this.url('Ping'),
      createHttpRequest('{}', headers, signal)).then((res) => {
      return buildResponse(res).then(_data => {
        return JsonDecode<PingResponse>(_data, 'PingResponse')
      })
    }, (error) => {
      throw WebrpcRequestFailedError.new({ cause: `fetch(): ${error instanceof Error ? error.message : String(error)}` })
    })
  }

  getUser = (req: GetUserRequest, headers?: object, signal?: AbortSignal): Promise<GetUserResponse> => {
    const _headers: Record<string, string> = { ...((headers || {}) as Record<string, string>) }
    if (req.authToken !== undefined && req.authToken !== null) {
      _headers["authToken"] = String(req.authToken)
    }
    return this.fetch(
      this.url('GetUser'),
      createHttpRequest(JsonEncode((() => { const {authToken: _h0, ..._body} = req; return _body})()), _headers, signal)).then((res) => {
      return buildResponse(res).then(_data => {
        return JsonDecode<GetUserResponse>(_data, 'GetUserResponse')
      })
    }, (error) => {
      throw WebrpcRequestFailedError.new({ cause: `fetch(): ${error instanceof Error ? error.message : String(error)}` })
    })
  }

  createUser = (req: CreateUserRequest, headers?: object, signal?: AbortSignal): Promise<CreateUserResponse> => {
    const _headers: Record<string, string> = { ...((headers || {}) as Record<string, string>) }
    if (req.authToken !== undefined && req.authToken !== null) {
      _headers["authToken"] = String(req.authToken)
    }
    if (req.role !== undefined && req.role !== null) {
      _headers["role"] = String(req.role)
    }
    return this.fetch(
      this.url('CreateUser'),
      createHttpRequest(JsonEncode((() => { const {authToken: _h0, role: _h1, ..._body} = req; return _body})()), _headers, signal)).then((res) => {
      return buildResponse(res).then(_data => {
        return JsonDecode<CreateUserResponse>(_data, 'CreateUserResponse')
      })
    }, (error) => {
      throw WebrpcRequestFailedError.new({ cause: `fetch(): ${error instanceof Error ? error.message : String(error)}` })
    })
  }

  deleteUser = (req: DeleteUserRequest, headers?: object, signal?: AbortSignal): Promise<void> => {
    const _headers: Record<string, string> = { ...((headers || {}) as Record<string, string>) }
    if (req.authToken !== undefined && req.authToken !== null) {
      _headers["authToken"] = String(req.authToken)
    }
    if (req.userID !== undefined && req.userID !== null) {
      _headers["userID"] = String(req.userID)
    }
    return this.fetch(
      this.url('DeleteUser'),
      createHttpRequest(JsonEncode({}), _headers, signal)).then((res) => {
      return buildResponse(res).then(_data => {
      })
    }, (error) => {
      throw WebrpcRequestFailedError.new({ cause: `fetch(): ${error instanceof Error ? error.message : String(error)}` })
    })
  }

}
// ...
export type Fetch = (input: RequestInfo, init?: RequestInit) => Promise<Response>
// ...
type WebrpcErrorParams = { name?: string, code?: number, message?: string, status?: number, cause?: string }

export class WebrpcError extends Error {
  code: number
  status: number

  constructor(error: WebrpcErrorParams = {}) {
    super(error.message)
    this.name = error.name || 'WebrpcEndpointError'
    this.code = typeof error.code === 'number' ? error.code : 0
    this.message = error.message || `endpoint error`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcError.prototype)
  }

  static new(payload: any): WebrpcError {
    return new this({ message: payload.message, code: payload.code, status: payload.status, cause: payload.cause })
  }
}
// ...
export class WebrpcRequestFailedError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcRequestFailed'
    this.code = typeof error.code === 'number' ? error.code : -1
    this.message = error.message || `request failed`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcRequestFailedError.prototype)
  }
}